Create LICENSE
Add License of GPL (GNU GENERAL PUBLIC LICENSE) V3 for this project. GPL license can be described such as below:

The GPL is often used for software projects and requires that, if developers distribute a modified version of the software, they do so under the same license. This means that if someone modifies your software and distributes the modified version, they must make the source code for that version available to others under the same terms.

The GPL also requires that users credit you as the original creator of the software and include a copy of the license with any distributions.